requirements.txt
Flask>=3.0.2
Werkzeug>=3.0.1
Pillow>=8.2.0
numpy>=1.20.3
SQLAlchemy>=1.4.15
Flask-Executor>=0.9.4
Flask-SQLAlchemy>=2.5.1
requests==2.31.0
pyyaml>=5.3.1
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.64.0
thop>=0.1.1
psutil
py-cpuinfo
kiwisolver>=1.3.1
fonttools>=4.22.0
cycler>=0.10
pyparsing>=2.3.1
packaging>=20.0
importlib-resources>=3.2.0
contourpy>=1.0.1
gitpython==3.1.42
setuptools==69.1.1
gitdb==4.0.11
smmap==5.0.1
dill==0.3.8
flask-cors==4.0.1
PyJWT==2.8.0
pymysql == 1.1.1
boto3 == 1.34.18
python-dotenv == 1.0.0